+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - eFuse write support (#962)
 - esp32p4rc1 support for esp32p4 chips with revision < 3.0 (#922)
+- Support for binaries with INIT_ARRAY sections, which is needed for esp32p4 support.
+- Add sha256 calculation to match esptool generated binaries
 
 ### Changed